pyproject.toml
[project]
name = "django-oauth2.0-template"
version = "0.1.0"
description = "Django authentication API using OAuth 2.0"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Thalysson Delano", email = "contato.thalyssondel@gmail.com" }
]

dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "django-allauth>=65.14.3",
    "dj-rest-auth>=7.1.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.9.0",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "cryptography>=44.0.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

exclude = ["**/migrations/*.py"]

[tool.ruff.lint]
# E, F = Erros básicos e PEP8
# B = Bugs comuns (Bugbear)
# S = Segurança (Bandit)
# I = Organização de Imports
select = ["E", "F", "B", "S", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # Permite o uso de 'assert' em arquivos de teste